Add working spritesheets for Walkablewith idle and walking animations (see Player.psd (not .png) for template). Fix animations to be more fluid / sync up with speed of player movement. Fix NPE caused by camera at game initialization

+++ src/edu/virginia/engine/Sys.java
@@ -19,7 +19,7 @@ public class Sys {
 	final static double VERTICAL_DRAG = 1.05;
 	final static double HORIZONTAL_SPEED = 1;
 	final static double GRAVITY = 1;
-	public final static double TILE_SIZE = 32;
+	public final static int TILE_SIZE = 32;
 	
 	//STATIC STORAGE
